Fix `Duration` and `Parameter` decoder
Fix #484

Since `HCursor#getOrElse` falls back the default value when `null` is given,
it causes unexpected bahavior. We now use `get` and `orElse` instead.
Because `Decoder` for `Option` works if the property is not given,
`Decoder` for `Duration` does not work with `Parameter` decoder. We use
`Either[Unit, Int]` decoder instead of `Option` decoder.